[refactor] 파일 디렉토리구조 변경.
