+++ components/atoms/button/Button.tsx
@@ -1,7 +1,7 @@
 import { VariantProps } from 'class-variance-authority';
 import { ButtonHTMLAttributes } from 'react';
 
-import { cn } from '@/utils/cn';
+import { cn } from '@/shared/utils/cn';
 
 import { buttonVariants } from './Button.style';
