+++ components/molecules/checkbox/Checkbox.tsx
@@ -4,7 +4,7 @@ import { forwardRef, InputHTMLAttributes, ReactNode, useId } from 'react';
 
 import { CheckboxIndicator } from '@/components/atoms/checkbox-indicator';
 import { Label } from '@/components/atoms/label';
-import { cn } from '@/utils/cn';
+import { cn } from '@/shared/utils/cn';
 
 import { checkboxVariants } from './Check.style';
