+++ components/molecules/multi-field/MultiField.tsx
@@ -2,7 +2,7 @@ import { ComponentPropsWithoutRef, useId } from 'react';
 
 import { Input } from '@/components/atoms/input';
 import { Label } from '@/components/atoms/label';
-import { cn } from '@/utils/cn';
+import { cn } from '@/shared/utils/cn';
 
 interface MultiFieldProps {
   id?: string;
